+++ docs/changelog.md
@@ -1,4 +1,5 @@
 dev
+- [bugfix] Prevent python error in case where package has no pipx metadata and advise user how to fix.
 
 0.15.4.0
 - [feature] `list` now has a new option `--include-injected` to show the injected packages in the main apps